+++ src/main/java/net/pincette/json/streams/Restart.java
@@ -1,27 +1,30 @@
 package net.pincette.json.streams;
 
 import static com.mongodb.client.model.Filters.eq;
+import static java.time.Instant.now;
 import static java.util.logging.Level.SEVERE;
 import static java.util.logging.Logger.getGlobal;
+import static net.pincette.jes.JsonFields.TIMESTAMP;
 import static net.pincette.json.streams.Application.APP_VERSION;
 import static net.pincette.json.streams.Common.APPLICATION_FIELD;
 import static net.pincette.json.streams.Common.getApplicationCollection;
-import static net.pincette.mongo.Collection.deleteOne;
+import static net.pincette.mongo.Collection.updateOne;
 import static net.pincette.util.Util.must;
 
-import com.mongodb.client.result.DeleteResult;
+import com.mongodb.client.model.Updates;
+import com.mongodb.client.result.UpdateResult;
 import java.util.function.Supplier;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.HelpCommand;
 import picocli.CommandLine.Option;
 
 @Command(
-    name = "delete",
+    name = "restart",
     version = APP_VERSION,
     mixinStandardHelpOptions = true,
     subcommands = {HelpCommand.class},
-    description = "Removes the application from the MongoDB collection, which causes it to stop.")
-class Delete implements Runnable {
+    description = "Restarts the application.")
+class Restart implements Runnable {
   private final Supplier<Context> contextSupplier;
 
   @Option(
@@ -32,15 +35,15 @@ class Delete implements Runnable {
 
   @Option(
       names = {"-c", "--collection"},
-      description = "The MongoDB collection the application is deleted from.")
+      description = "The MongoDB collection the application is restarted in.")
   private String collection;
 
-  Delete(final Supplier<Context> contextSupplier) {
+  Restart(final Supplier<Context> contextSupplier) {
     this.contextSupplier = contextSupplier;
   }
 
-  private DeleteResult logExists(final DeleteResult result) {
-    if (result.getDeletedCount() != 1) {
+  private UpdateResult logExists(final UpdateResult result) {
+    if (result.getModifiedCount() != 1) {
       getGlobal().log(SEVERE, "The application {0} doesn''t exist.", application);
     }
 
@@ -50,10 +53,11 @@ class Delete implements Runnable {
   public void run() {
     final var context = contextSupplier.get();
 
-    deleteOne(
+    updateOne(
             context.database.getCollection(getApplicationCollection(collection, context)),
-            eq(APPLICATION_FIELD, application))
-        .thenApply(result -> must(result, DeleteResult::wasAcknowledged))
+            eq(APPLICATION_FIELD, application),
+            Updates.set(TIMESTAMP, now().toString()))
+        .thenApply(result -> must(result, UpdateResult::wasAcknowledged))
         .thenApply(this::logExists)
         .toCompletableFuture()
         .join();